refactor: メニューバーの修正
