+++ quasar.conf.js
@@ -84,6 +84,7 @@ module.exports = function (/* ctx */) {
     framework: {
       iconSet: 'material-icons', // Quasar icon set
       lang: 'en-us', // Quasar language pack
+      cssAddon:true,
       config: {
         loading:{}
       },
