+++ src/pages/UserList.vue
@@ -1,39 +1,45 @@
 <template>
-  <q-page class="flex flex-center column">
-
-    <h2>UserList</h2>
-    <div class="q-pa-md row flex-start">
-      <div class="col-3 q-px-sm q-my-sm registered-user" v-for="(user, index) in userList.registered" :key="user.name"
-        @click="(e) => propClickUser(e)" :data-key="index">
-        <q-card>
-          <img :src="user.img">
-          <q-card-section>
-            <div class="balloon">一言コメント・PR</div>
-            <div class="text-h6">{{user.name}}</div>
-            <div class="text-subtitle2">
-              age:
-              <span v-if="user.birth === null">---</span>
-              <span v-else>{{user.birth}}</span>
-            </div>
-          </q-card-section>
-        </q-card>
+  <q-page>
+    <div class="" v-if="nowLoading"></div>
+    <div class="flex flex-center column" v-else>
+      <h2>UserList</h2>
+      <div class="q-pa-md row flex-start" :class="{hidden: nowLoading}">
+        <div class="col-12 col-sm-4 col-md-3 q-px-sm q-my-sm registered-user"
+          v-for="(user, index) in userList.registered" :key="user.name" @click="(e) => propClickUser(e)"
+          :data-key="index">
+          <q-card class="row">
+            <img :src="user.img" class="col-5 col-sm-12 q-px-md q-pt-md">
+            <q-card-section class="order-sm-last col-7 col-sm-12">
+              <div class="text-h6">{{user.name}}</div>
+              <div class="text-subtitle2">
+                age:
+                <span v-if="user.birth === null">---</span>
+                <span v-else>{{user.birth}}</span>
+              </div>
+            </q-card-section>
+            <q-card-section class="col">
+              <div class="balloon">一言コメント・PR</div>
+            </q-card-section>
+          </q-card>
+        </div>
+        <div class="col-12 col-sm-4 col-md-3 q-px-sm q-my-sm sample-user" v-for="(sample, index) in userList.sample"
+          :key="sample.id.value" @click="(e) => propClickUser(e)" :data-key="index">
+          <q-card class="row">
+            <img :src="sample.picture.large" class="col-5 col-sm-12 q-px-md q-pt-md">
+            <q-card-section class="order-sm-last col-7 col-sm-12">
+              <div class="text-h6">{{sample.name.first}}(Sample)</div>
+              <div class="text-subtitle2">age: {{sample.dob.age}}</div>
+            </q-card-section>
+            <q-card-section class="col">
+              <div class="balloon">一言コメント・PR</div>
+            </q-card-section>
+          </q-card>
+        </div>
+        <q-dialog v-model="openDialog">
+          <UserDetail :detail="clickedUserInfo" />
+        </q-dialog>
       </div>
-      <div class="col-3 q-px-sm q-my-sm sample-user" v-for="(sample, index) in userList.sample" :key="sample.id.value"
-        @click="(e) => propClickUser(e)" :data-key="index">
-        <q-card>
-          <img :src="sample.picture.large">
-          <q-card-section>
-            <div class="balloon">一言コメント・PR</div>
-            <div class="text-h6">{{sample.name.first}}(Sample)</div>
-            <div class="text-subtitle2">age: {{sample.dob.age}}</div>
-          </q-card-section>
-        </q-card>
-      </div>
-      <q-dialog v-model="openDialog">
-        <UserDetail :detail="clickedUserInfo" />
-      </q-dialog>
     </div>
-
   </q-page>
 </template>
 
@@ -54,6 +60,7 @@
     name: 'UserList',
     data() {
       return {
+        nowLoading: true,
         userList: {},
         clickedUserInfo: {},
         openDialog: false,
@@ -71,14 +78,9 @@
         const q = query(collection(db, "users"), where("uid", "!=", loginUser.uid));
         const queryUser = await getDocs(q);
         const arr = [];
-        console.log(queryUser)
         queryUser.forEach(doc => arr.push(doc.data()));
 
-        let response = this.getUserList.sample;
-        console.log(this.getUserList)
-        if (response.length === 0) {
-          response = await fetch(api.url, api.option).then(res => res.json());
-        }
+        const response = await fetch(api.url, api.option).then(res => res.json());
 
         this.setSample(response.results);
         this.setRegistered(arr);
@@ -107,21 +109,29 @@
 
     created() {
       this.$q.loading.show();
-      const userList = JSON.parse(localStorage.getItem("userList"));
-      const storeUserList = this.getUserList;
-      if (userList != null) {
-        // ユーザーリストが保存されている時は内容をストアで受け取る
-        this.setRegistered(userList.registered);
-        this.setSample(userList.sample);
-      } else if (storeUserList.registered.length === 0 || storeUserList.sample.length === 0) {
-        // ストアにユーザーリストが無い場合は読み込みを実行する
+      const list = JSON.parse(localStorage.getItem("userList"));
+      const storelist = this.getUserList;
+      if (list != null) {
+        // ローカルにリストが保存されている時（ページ更新時）
+        this.setRegistered(list.registered);
+        this.setSample(list.sample);
+      } else if (storelist.registered.length === 0 || storelist.sample.length === 0) {
+        // ローカルにもストアにもリストが無い場合（ログイン時）
         this.getUser();
       }
       this.userList = this.getUserList;
-      this.$q.loading.hide();
-      console.log(this.userList)
       // 更新する直前でユーザーリストを保存する
       window.addEventListener("beforeunload", this.holdUsersAtReload);
+    },
+    mounted() {
+      const storelist = this.getUserList;
+      this.$nextTick(() => {
+        this.timer = setTimeout(() => {
+          this.$q.loading.hide()
+          this.timer = void 0
+          this.nowLoading = false;
+        }, 1500)
+      })
     },
   }
   /*
